requirements.txt
python-dotenv>=0.19.0,<2.0.0
web3>=5.31.3,<7.0.0
websockets>=10.0.0,<12.0.0
requests>=2.25.0,<3.0.0
py-clob-client @ git+https://github.com/Polymarket/py-clob-client.git
python-dateutil>=2.8.0,<3.0.0
aiohttp>=3.8.0,<4.0.0
asyncio>=3.4.3,<4.0.0 